+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/base/MentorJob.kt
@@ -1,14 +1,12 @@
 package com.sourceplusplus.mentor.base
 
-import com.sourceplusplus.mentor.base.MentorJob.TaskContext
 import com.sourceplusplus.protocol.advice.AdviceListener
 import com.sourceplusplus.protocol.advice.ArtifactAdvice
 import io.vertx.core.Vertx
 import org.slf4j.LoggerFactory
-import java.util.*
 
 /**
- * Primarily used to propagate a persistent context ([TaskContext]) between [MentorTask]s.
+ * Keeps track of the order and current position of [MentorTask]s which require processing.
  *
  * @since 0.0.1
  * @author [Brandon Fergerson](mailto:bfergerson@apache.org)
@@ -23,7 +21,7 @@ abstract class MentorJob {
     var config: MentorJobConfig = MentorJobConfig()
         private set
     abstract val tasks: List<MentorTask>
-    val context = TaskContext()
+    val context = MentorTaskContext()
     private var currentTask = -1
     private var complete: Boolean = false
     private val listeners: MutableList<MentorJobListener> = mutableListOf()
@@ -72,48 +70,4 @@ abstract class MentorJob {
     }
 
     override fun toString(): String = "${javaClass.simpleName}@${System.identityHashCode(this)}"
-
-    class TaskContext {
-        private val cache: IdentityHashMap<ContextKey<*>, Any> = IdentityHashMap()
-
-        fun <T> put(key: ContextKey<T>, value: T) {
-            cache[key] = value!!
-        }
-
-        fun <T> get(key: ContextKey<T>): T {
-            return cache[key] as T
-        }
-
-        internal fun clear() {
-            cache.clear()
-        }
-
-        fun copyOutputContext(copyJob: MentorJob, copyTask: MentorTask) {
-            copyOutputContext(copyJob.context, copyTask)
-        }
-
-        fun copyOutputContext(copyContext: TaskContext, copyTask: MentorTask) {
-            copyTask.outputContextKeys.forEach {
-                cache[it] = copyContext.get(it)
-            }
-        }
-
-        fun copyFullContext(copyJob: MentorJob, copyTask: MentorTask) {
-            copyFullContext(copyJob.context, copyTask)
-        }
-
-        fun copyFullContext(copyContext: TaskContext, copyTask: MentorTask) {
-            copyTask.inputContextKeys.forEach {
-                cache[it] = copyContext.get(it)
-            }
-            copyTask.outputContextKeys.forEach {
-                cache[it] = copyContext.get(it)
-            }
-        }
-    }
-
-    @Suppress("unused")
-    class ContextKey<T>(private val name: String) {
-        override fun toString(): String = name
-    }
 }
