+++ mentor/src/main/kotlin/com/sourceplusplus/mentor/impl/task/analyze/DetermineThrowableLocation.kt
@@ -1,7 +1,7 @@
 package com.sourceplusplus.mentor.impl.task.analyze
 
+import com.sourceplusplus.mentor.base.ContextKey
 import com.sourceplusplus.mentor.base.MentorJob
-import com.sourceplusplus.mentor.base.MentorJob.ContextKey
 import com.sourceplusplus.mentor.base.MentorTask
 import com.sourceplusplus.protocol.advice.informative.ActiveExceptionAdvice
 import com.sourceplusplus.protocol.artifact.ArtifactLocation
